Cleaned the code up a little
