+++ CD/Makefile
@@ -1,0 +1,2 @@
+cd.exe: src/main.cpp src/cd.cpp src/cd.hpp src/signedarray.hpp
+	g++ -std=c++17 src/main.cpp src/cd.cpp -o cd.exe
